src/etl/main.py
import polars as pl
import psycopg2
import boto3
import os
from transform import transform_data

def get_instance_id():
    """Get the current EC2 instance ID"""
    response = requests.get('http://169.254.169.254/latest/meta-data/instance-id')
    return response.text

def run_etl():
    print("Starting ETL process...")
    
    try:
        # Extract from Redshift
        # ... extraction logic ...
        
        # Transform
        # ... transformation logic ...
        
        # Load to PostgreSQL
        # ... loading logic ...
        
        print("ETL process completed")
        
        # Send completion notification
        sns = boto3.client('sns')
        instance_id = get_instance_id()
        
        sns.publish(
            TopicArn=os.environ['COMPLETION_TOPIC_ARN'],
            Message=json.dumps({
                'status': 'completed',
                'instance_id': instance_id
            })
        )
        
    except Exception as e:
        # Send failure notification
        sns.publish(
            TopicArn=os.environ['COMPLETION_TOPIC_ARN'],
            Message=json.dumps({
                'status': 'failed',
                'instance_id': instance_id,
                'error': str(e)
            })
        )
        raise e

if __name__ == "__main__":
    run_etl()