src/etl/requirements.txt
polars==0.19.3
psycopg2-binary==2.9.9
boto3==1.28.44